Add role-wise checks
